ref(repo): move everything to `lib` dir to keep private items private
