+++ src/lib/get_btc_private_key.rs
@@ -1,5 +1,5 @@
 use std::process::Command;
-use crate::{
+use crate::lib::{
     state::State,
     types::Result,
     errors::AppError,
