+++ src/lib/get_utxo_json_string.rs
@@ -1,4 +1,4 @@
-use crate::{
+use crate::lib::{
     state::State,
     types::Result,
     errors::AppError,
