+++ src/lib/make_online_transaction.rs
@@ -1,4 +1,4 @@
-use crate::{
+use crate::lib::{
     state::State,
     types::Result,
     get_cli_args::CliArgs,
