+++ src/lib/mod.rs
@@ -1,0 +1,27 @@
+pub(crate) mod state;
+pub(crate) mod types;
+pub(crate) mod utils;
+pub(crate) mod base58;
+pub(crate) mod errors;
+pub(crate) mod get_utxos;
+pub(crate) mod create_tx;
+pub(crate) mod constants;
+pub(crate) mod utxo_codec;
+pub(crate) mod test_utils;
+pub(crate) mod usage_info;
+pub(crate) mod save_output;
+pub(crate) mod get_cli_args;
+pub(crate) mod get_utxos_info;
+pub(crate) mod btc_transaction;
+pub(crate) mod btc_private_key;
+pub(crate) mod initialize_logger;
+pub(crate) mod create_op_return_tx;
+pub(crate) mod get_btc_private_key;
+pub(crate) mod get_utxo_json_string;
+pub(crate) mod get_utxos_for_address;
+pub(crate) mod make_online_transaction;
+pub(crate) mod make_offline_transaction;
+pub(crate) mod get_btc_txs_from_utxos_info;
+pub(crate) mod extract_utxos_from_utxo_info;
+pub(crate) mod make_online_op_return_transaction;
+pub(crate) mod make_offline_op_return_transaction;
