+++ src/lib/utils.rs
@@ -4,7 +4,7 @@ use rand::{
     RngCore,
     thread_rng,
 };
-use crate::{
+use crate::lib::{
     state::State,
     errors::AppError,
     types::{
