+++ src/lib/utxo_codec.rs
@@ -1,4 +1,4 @@
-use crate::{
+use crate::lib::{
     state::State,
     types::{
         Result,
