+++ src/main.rs
@@ -1,38 +1,11 @@
 #![feature(try_trait)]
 #![allow(clippy::match_bool)]
 
-pub mod state;
-pub mod types;
-pub mod utils;
-pub mod base58;
-pub mod errors;
-pub mod get_utxos;
-pub mod create_tx;
-pub mod constants;
-pub mod utxo_codec;
-pub mod test_utils;
-pub mod usage_info;
-pub mod save_output;
-pub mod get_cli_args;
-pub mod get_utxos_info;
-pub mod btc_transaction;
-pub mod btc_private_key;
-pub mod initialize_logger;
-pub mod create_op_return_tx;
-pub mod get_btc_private_key;
-pub mod get_utxo_json_string;
-pub mod get_utxos_for_address;
-pub mod make_online_transaction;
-pub mod make_offline_transaction;
-pub mod get_btc_txs_from_utxos_info;
-pub mod extract_utxos_from_utxo_info;
-pub mod make_online_op_return_transaction;
-pub mod make_offline_op_return_transaction;
-
+mod lib;
 #[macro_use] extern crate log;
 #[macro_use] extern crate serde_derive;
 
-use crate::{
+use crate::lib::{
     types::Result,
     errors::AppError,
     get_utxos::get_utxos,
